Move RemindEnterPasswordModal to components dir
